refactor: added double border for fterm to easier identify
Former-commit-id: ec6a797f74023a60066e1a5cffc94b31ac0d76cc
Former-commit-id: fdf7fa8f01ca1b95158592ed0483e3cef2155e7d

+++ lua/plugins/fterm.lua
@@ -9,7 +9,7 @@ require('FTerm').setup {
 		width = 0.9,
 	},
 
-	border = 'single'
+	border = 'double'
 }
 
 local option = {noremap=true, silent=true}
